+++ sdk/network/arm-network/src/models/index.ts
@@ -11393,6 +11393,8 @@ export enum KnownApplicationGatewaySkuName {
   WAFLarge = "WAF_Large",
   /** StandardV2 */
   StandardV2 = "Standard_v2",
+  /** StandardBasic */
+  StandardBasic = "Standard_Basic",
   /** WAFV2 */
   WAFV2 = "WAF_v2"
 }
@@ -11408,6 +11410,7 @@ export enum KnownApplicationGatewaySkuName {
  * **WAF_Medium** \
  * **WAF_Large** \
  * **Standard_v2** \
+ * **Standard_Basic** \
  * **WAF_v2**
  */
 export type ApplicationGatewaySkuName = string;
@@ -11421,7 +11424,9 @@ export enum KnownApplicationGatewayTier {
   /** StandardV2 */
   StandardV2 = "Standard_v2",
   /** WAFV2 */
-  WAFV2 = "WAF_v2"
+  WAFV2 = "WAF_v2",
+  /** StandardBasic */
+  StandardBasic = "Standard_Basic"
 }
 
 /**
@@ -11432,7 +11437,8 @@ export enum KnownApplicationGatewayTier {
  * **Standard** \
  * **WAF** \
  * **Standard_v2** \
- * **WAF_v2**
+ * **WAF_v2** \
+ * **Standard_Basic**
  */
 export type ApplicationGatewayTier = string;
 
@@ -16817,10 +16823,7 @@ export type NetworkInterfacesListVirtualMachineScaleSetNetworkInterfacesNextResp
 
 /** Optional parameters. */
 export interface NetworkInterfacesListVirtualMachineScaleSetIpConfigurationsNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** Expands referenced resources. */
-  expand?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listVirtualMachineScaleSetIpConfigurationsNext operation. */
 export type NetworkInterfacesListVirtualMachineScaleSetIpConfigurationsNextResponse = NetworkInterfaceIPConfigurationListResult;
@@ -18559,24 +18562,14 @@ export type NetworkManagersListResponse = NetworkManagerListResult;
 
 /** Optional parameters. */
 export interface NetworkManagersListBySubscriptionNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** An optional query parameter which specifies the maximum number of records to be returned by the server. */
-  top?: number;
-  /** SkipToken is only used if a previous operation returned a partial result. If a previous response contains a nextLink element, the value of the nextLink element will include a skipToken parameter that specifies a starting point to use for subsequent calls. */
-  skipToken?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listBySubscriptionNext operation. */
 export type NetworkManagersListBySubscriptionNextResponse = NetworkManagerListResult;
 
 /** Optional parameters. */
 export interface NetworkManagersListNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** An optional query parameter which specifies the maximum number of records to be returned by the server. */
-  top?: number;
-  /** SkipToken is only used if a previous operation returned a partial result. If a previous response contains a nextLink element, the value of the nextLink element will include a skipToken parameter that specifies a starting point to use for subsequent calls. */
-  skipToken?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listNext operation. */
 export type NetworkManagersListNextResponse = NetworkManagerListResult;
@@ -18635,12 +18628,7 @@ export type SubscriptionNetworkManagerConnectionsListResponse = NetworkManagerCo
 
 /** Optional parameters. */
 export interface SubscriptionNetworkManagerConnectionsListNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** An optional query parameter which specifies the maximum number of records to be returned by the server. */
-  top?: number;
-  /** SkipToken is only used if a previous operation returned a partial result. If a previous response contains a nextLink element, the value of the nextLink element will include a skipToken parameter that specifies a starting point to use for subsequent calls. */
-  skipToken?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listNext operation. */
 export type SubscriptionNetworkManagerConnectionsListNextResponse = NetworkManagerConnectionListResult;
@@ -18677,12 +18665,7 @@ export type ManagementGroupNetworkManagerConnectionsListResponse = NetworkManage
 
 /** Optional parameters. */
 export interface ManagementGroupNetworkManagerConnectionsListNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** An optional query parameter which specifies the maximum number of records to be returned by the server. */
-  top?: number;
-  /** SkipToken is only used if a previous operation returned a partial result. If a previous response contains a nextLink element, the value of the nextLink element will include a skipToken parameter that specifies a starting point to use for subsequent calls. */
-  skipToken?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listNext operation. */
 export type ManagementGroupNetworkManagerConnectionsListNextResponse = NetworkManagerConnectionListResult;
@@ -18726,12 +18709,7 @@ export type ConnectivityConfigurationsListResponse = ConnectivityConfigurationLi
 
 /** Optional parameters. */
 export interface ConnectivityConfigurationsListNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** An optional query parameter which specifies the maximum number of records to be returned by the server. */
-  top?: number;
-  /** SkipToken is only used if a previous operation returned a partial result. If a previous response contains a nextLink element, the value of the nextLink element will include a skipToken parameter that specifies a starting point to use for subsequent calls. */
-  skipToken?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listNext operation. */
 export type ConnectivityConfigurationsListNextResponse = ConnectivityConfigurationListResult;
@@ -18779,12 +18757,7 @@ export type NetworkGroupsListResponse = NetworkGroupListResult;
 
 /** Optional parameters. */
 export interface NetworkGroupsListNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** An optional query parameter which specifies the maximum number of records to be returned by the server. */
-  top?: number;
-  /** SkipToken is only used if a previous operation returned a partial result. If a previous response contains a nextLink element, the value of the nextLink element will include a skipToken parameter that specifies a starting point to use for subsequent calls. */
-  skipToken?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listNext operation. */
 export type NetworkGroupsListNextResponse = NetworkGroupListResult;
@@ -18821,12 +18794,7 @@ export type StaticMembersListResponse = StaticMemberListResult;
 
 /** Optional parameters. */
 export interface StaticMembersListNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** An optional query parameter which specifies the maximum number of records to be returned by the server. */
-  top?: number;
-  /** SkipToken is only used if a previous operation returned a partial result. If a previous response contains a nextLink element, the value of the nextLink element will include a skipToken parameter that specifies a starting point to use for subsequent calls. */
-  skipToken?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listNext operation. */
 export type StaticMembersListNextResponse = StaticMemberListResult;
@@ -18863,12 +18831,7 @@ export type ScopeConnectionsListResponse = ScopeConnectionListResult;
 
 /** Optional parameters. */
 export interface ScopeConnectionsListNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** An optional query parameter which specifies the maximum number of records to be returned by the server. */
-  top?: number;
-  /** SkipToken is only used if a previous operation returned a partial result. If a previous response contains a nextLink element, the value of the nextLink element will include a skipToken parameter that specifies a starting point to use for subsequent calls. */
-  skipToken?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listNext operation. */
 export type ScopeConnectionsListNextResponse = ScopeConnectionListResult;
@@ -18912,12 +18875,7 @@ export interface SecurityAdminConfigurationsDeleteOptionalParams
 
 /** Optional parameters. */
 export interface SecurityAdminConfigurationsListNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** An optional query parameter which specifies the maximum number of records to be returned by the server. */
-  top?: number;
-  /** SkipToken is only used if a previous operation returned a partial result. If a previous response contains a nextLink element, the value of the nextLink element will include a skipToken parameter that specifies a starting point to use for subsequent calls. */
-  skipToken?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listNext operation. */
 export type SecurityAdminConfigurationsListNextResponse = SecurityAdminConfigurationListResult;
@@ -18961,12 +18919,7 @@ export interface AdminRuleCollectionsDeleteOptionalParams
 
 /** Optional parameters. */
 export interface AdminRuleCollectionsListNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** An optional query parameter which specifies the maximum number of records to be returned by the server. */
-  top?: number;
-  /** SkipToken is only used if a previous operation returned a partial result. If a previous response contains a nextLink element, the value of the nextLink element will include a skipToken parameter that specifies a starting point to use for subsequent calls. */
-  skipToken?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listNext operation. */
 export type AdminRuleCollectionsListNextResponse = AdminRuleCollectionListResult;
@@ -19010,12 +18963,7 @@ export interface AdminRulesDeleteOptionalParams
 
 /** Optional parameters. */
 export interface AdminRulesListNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** An optional query parameter which specifies the maximum number of records to be returned by the server. */
-  top?: number;
-  /** SkipToken is only used if a previous operation returned a partial result. If a previous response contains a nextLink element, the value of the nextLink element will include a skipToken parameter that specifies a starting point to use for subsequent calls. */
-  skipToken?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listNext operation. */
 export type AdminRulesListNextResponse = AdminRuleListResult;
@@ -20493,12 +20441,7 @@ export type ServiceTagInformationListResponse = ServiceTagInformationListResult;
 
 /** Optional parameters. */
 export interface ServiceTagInformationListNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** Do not return address prefixes for the tag(s). */
-  noAddressPrefixes?: boolean;
-  /** Return tag information for a particular tag. */
-  tagName?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listNext operation. */
 export type ServiceTagInformationListNextResponse = ServiceTagInformationListResult;
@@ -20621,12 +20564,7 @@ export type VirtualNetworksListUsageNextResponse = VirtualNetworkListUsageResult
 
 /** Optional parameters. */
 export interface VirtualNetworksListDdosProtectionStatusNextOptionalParams
-  extends coreClient.OperationOptions {
-  /** The max number of ip addresses to return. */
-  top?: number;
-  /** The skipToken that is given with nextLink. */
-  skipToken?: string;
-}
+  extends coreClient.OperationOptions {}
 
 /** Contains response data for the listDdosProtectionStatusNext operation. */
 export type VirtualNetworksListDdosProtectionStatusNextResponse = VirtualNetworkDdosProtectionStatusResult;